+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "pcd-viewer" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.5] - 2022-11-12
 
-- Initial release
+### Added
+
+- Added adjustable point style and point size.
+- Added reset the visual center to the coordinate origin or point cloud center.
+
+### Changed
+
+- Changed vscode dependency to `^1.57.0` to pass data useing `TypedArray` without serialization.